fix(clippy): clear the workspace of lints
Seven lints, all mechanical:

- `Config::{load,write,load_or_create}` took `&PathBuf` where `&Path` does
  (`ptr_arg`). Callers are unaffected — deref coercion already passed `&PathBuf`
  through. `load_or_create` only surfaced once `load` was fixed, since passing
  its `&PathBuf` straight through had been masking it.
- A redundant `k as u32` inside `(0..n as u32)`, where `k` is already `u32`.
- `1 * 3 * 6` in a stroke-mesh assertion; the "1 segment" it documented is
  already stated in the comment above it.
- Two `VisSettings::default()`-then-assign blocks folded into struct literals,
  and two `&vec![..]` test arguments made `&[..]`.

+++ crates/bava/src/vis/physics.rs
@@ -2600,9 +2600,7 @@ mod tests {
         app.insert_resource(Planet {
             radii: radii.clone(),
             prev: radii, // steady → no expansion fling; escape must come from unstick
-            indices: (0..n as u32)
-                .map(|k| [k, (k as u32 + 1) % n as u32])
-                .collect(),
+            indices: (0..n as u32).map(|k| [k, (k + 1) % n as u32]).collect(),
         });
 
         // Clear Bevy's zero-dt first frame (planet_forces no-ops on dt==0) before
